refactor: cambio de administracion de memoria
